Lodha Bund Garden Pune
Lodha Bund Garden Pune offers apartments in a varied budget range. These units are a perfect mixture of luxury and style, specifically designed to suit your requirements and amenities. Built with care using quality materials and modern conveniences, these spacious flats come with all necessary facilities like a lift, parking space, and water supply.

It is a 1, 2 & 3 BHK Premium Residences in this project, which have separate balconies and stairways for each floor. This housing society is now ready to be called home as families have started moving in.

The open living space of the home is intended with many little pockets of activity, as well as generous meeting spaces and points of view. The centralized pillar supports the circular volume, and fluid separations create a sense of flow while providing privacy to each space.

For More Details :
Visit - https://lodha.developerprojects.in/residential/pune/lodha-bund-garden/